Federated Learning/Cifar10 - Federated Learning/requirements.txt
tensorflow
flwr
hydra-core